+++ src/qudi/interface/scanning_probe_interface.py
@@ -22,7 +22,7 @@ If not, see <https://www.gnu.org/licenses/>.
 
 from enum import auto, Flag
 from dataclasses import dataclass, field, asdict, replace
-from typing import Optional, Tuple, Dict
+from typing import Optional, Tuple, Dict, Any
 import datetime
 import numpy as np
 from abc import abstractmethod
@@ -271,6 +271,7 @@ class ScanData:
     _data: Optional[Tuple[np.ndarray, ...]] = None
     # TODO: Automatic interpolation onto rectangular grid needs to be implemented (for position feedback HW)
     _position_data: Optional[Tuple[np.ndarray, ...]] = None
+    coord_transform_info: Dict[str, Any] = field(default_factory=dict)
 
     @classmethod
     def from_constraints(cls, settings: ScanSettings, constraints: ScanConstraints, **kwargs):
@@ -394,6 +395,27 @@ class ScanningProbeInterface(Base):
 
     A scanner device is hardware that can move multiple axes.
     """
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._coordinate_transform = None
+        self._coordinate_transform_matrix = None
+
+    def coordinate_transform(self, val, inverse=False):
+        if self._coordinate_transform is None:
+            return val
+        return self._coordinate_transform(val, inverse)
+
+    def set_coordinate_transform(self, transform_func, transform_matrix=None):
+        if transform_func is not None:
+            raise ValueError('Coordinate transformation not supported by scanning hardware.')
+
+    @property
+    def coordinate_transform_enabled(self):
+        return self._coordinate_transform is not None
+
+    @property
+    def supports_coordinate_transform(self):
+        return isinstance(self, CoordinateTransformMixin)
 
     @property
     @abstractmethod
@@ -528,3 +550,97 @@ class ScanningProbeInterface(Base):
         TODO: document what this should to differently than stop_scan.
         """
         pass
+
+    def _expand_coordinate(self, coord):
+        """
+        Expand coord dict to all scanner dimensions, setting missing axes to current scanner target.
+        """
+
+        scanner_axes = self.constraints.axes
+        current_target = self.get_target()
+        len_coord = 0
+        axes_unused = scanner_axes.keys()
+
+        if coord:
+            len_coord = np.asarray((list(coord.values())[0])).size
+            axes_unused = [ax for ax in scanner_axes.keys() if ax not in coord.keys()]
+        coord_unused = {}
+
+        for ax in axes_unused:
+            target_coord = current_target[ax]
+            coords = np.ones(len_coord)*target_coord if len_coord > 1 else target_coord
+            coord_unused[ax] = coords
+
+        coord.update(coord_unused)
+
+        return coord
+
+
+class CoordinateTransformMixin(ScanningProbeInterface):
+    """ Can be used by concrete hardware modules to facilitate coordinate transformation, except
+    for performing scans.
+    The transformation for scanning can be either implemented in the base or in the mixed hardware
+    module.
+
+    Usage:
+        MyTransformationScanner(CoordinateTransformMixin, MyScanner):
+            pass
+    """
+    def set_coordinate_transform(self, transform_func, transform_matrix=None):
+        # ToDo: Proper sanity checking here, e.g. function signature etc.
+        if transform_func is not None and not callable(transform_func):
+            raise ValueError('Coordinate transformation function must be callable with '
+                             'signature "coordinate_transform(value, inverse=False)"')
+        self._coordinate_transform = transform_func
+        self._coordinate_transform_matrix = transform_matrix
+
+    def move_absolute(self, position, velocity=None, blocking=False):
+        new_pos_bare = super().move_absolute(self.coordinate_transform(position), velocity, blocking)
+        return self.coordinate_transform(new_pos_bare, inverse=True)
+
+    def move_relative(self, distance, velocity=None, blocking=False):
+        new_pos_bare = super().move_relative(self.coordinate_transform(distance), velocity, blocking)
+        return self.coordinate_transform(new_pos_bare, inverse=True)
+
+    def get_target(self):
+        return self.coordinate_transform(super().get_target(), inverse=True)
+
+    def get_position(self):
+        return self.coordinate_transform(super().get_position(), inverse=True)
+
+    def _calc_matr_2_tiltangle(self):
+        """
+        Calculates the tilt angle in radians of a given rotation matrix.
+        Formula from https://en.wikipedia.org/wiki/Rotation_matrix
+        """
+
+        rotation_matrix = self._coordinate_transform_matrix.matrix[0:3,0:3]
+        trace = np.trace(rotation_matrix)
+        tilt_angle_abs = np.arccos((trace-1)/2)
+        return tilt_angle_abs
+
+    def get_scan_data(self):
+        scan_data = super().get_scan_data()
+        if scan_data:
+            if self.coordinate_transform_enabled:
+                scan_data.coord_transform_info = self._get_coord_transform_info()
+            else:
+                scan_data.coord_transform_info = {}
+        return scan_data
+
+    def get_back_scan_data(self):
+        scan_data = super().get_back_scan_data()
+        if scan_data:
+            if self.coordinate_transform_enabled:
+                scan_data.coord_transform_info = self._get_coord_transform_info()
+            else:
+                scan_data.coord_transform_info = {}
+        return scan_data
+
+    def _get_coord_transform_info(self) -> Dict[str, Any]:
+        rad_2_deg = lambda phi: phi / np.pi * 180
+        return {
+            'transform_matrix': self._coordinate_transform_matrix.matrix,
+            'tilt_angle (deg)': rad_2_deg(self._calc_matr_2_tiltangle()),
+            'translation': self._coordinate_transform_matrix.matrix[0:3, -1]
+        }